Add an out-of-context deployment container
This creates the context where:

- There are many systems and containers
- All containers are linked to the same deployment node
- The deployment view is only in scope for one of the systems

The previous context had a single system.

This setup breaks the current tests, as the deployment view generated by
the PlantUML writer includes the containers added to the deployment node
but not in scope for the system.

This change renders the same diagram in the Structurizr web UI.

TL;DR: PlantUML deployment diagrams render "out of scope" containers.